agente-ia-odonto/services/requirements.txt
# services/requirements.txt - ETAPA 5

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# HTTP Client
httpx==0.25.1
aiohttp==3.9.0

# LLM Providers
anthropic==0.7.7
openai==1.3.7
google-generativeai==0.3.0
groq==0.3.0

# Google Calendar - NOVO NA ETAPA 5
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.111.0

# Timezone - NOVO NA ETAPA 5
pytz==2023.3.post1

# Utils
pydantic==2.5.2
pyyaml==6.0.1

# Dev/Test
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0